Factor out `LabelLike` to share common label styles
